refactor(workspace): isolate refresh coordination
Problem:
The review fixes left refresh coordination and mutation tests embedded in already oversized owner files, and one source-contract test assumed the fallback lived directly in App.tsx.

Root cause:
Fast-path mutation delivery and focus reconciliation were added at their call sites before the final repository-size audit, coupling the behavior to presentation files and historical test layout.

Fix:
Extract host mutation contracts, mutation finalization, focus reconciliation, and focused tests into dedicated modules. Keep resource flags on the host-only signal, preserve provider wire omission, and teach the source contract about the extracted hook without expanding any lint baseline.

Verification:
- go test ./... and go vet ./...
- desktop go test ./... and go vet ./...
- focused Go race suites for agent, event, eventwire, boot, and desktop workspace hub
- pnpm test:workspace, pnpm test:all, and pnpm build
- repolint, release workflow contracts, formatting, bundle budgets, and diff checks
