+++ desktop/frontend/src/__tests__/app-chrome-tabs.test.ts
@@ -7,7 +7,7 @@ import { createBoundedRefreshCoordinator, sameTabMetaLists, shouldRefreshTabMeta
 import type { TabMeta } from "../lib/types";
 
 const testDir = dirname(fileURLToPath(import.meta.url));
-const appSource = readFileSync(resolve(testDir, "../App.tsx"), "utf8");
+const appSource = readFileSync(resolve(testDir, "../App.tsx"), "utf8"), workspaceFocusSource = readFileSync(resolve(testDir, "../lib/useWorkspaceFocusReconciliation.ts"), "utf8");
 const appChromeSource = readFileSync(resolve(testDir, "../components/AppChrome.tsx"), "utf8");
 const commandPaletteSource = readFileSync(resolve(testDir, "../components/CommandPalette.tsx"), "utf8");
 const projectTreeSource = readFileSync(resolve(testDir, "../components/ProjectTree.tsx"), "utf8");
@@ -227,9 +227,9 @@ ok(!shouldRefreshTabMetaForEvent("text_delta"), "stream deltas do not trigger ta
 
 ok(
   !appSource.includes("setInterval(() => void refreshTabMetas(), 2000)") &&
-    appSource.includes('document.addEventListener("visibilitychange", onVisibilityChange)') &&
+    workspaceFocusSource.includes('document.addEventListener("visibilitychange", onVisibilityChange)') &&
     appSource.includes("createBoundedRefreshCoordinator<TabMeta[]>(TAB_META_MAX_IN_FLIGHT)") &&
-    appSource.includes("void refreshTabMetas();\n        schedule();"),
+    workspaceFocusSource.includes("void refreshTabMetas();\n        schedule();"),
   "tab metadata refresh is event-driven with a visibility-aware fallback",
 );
