+++ internal/agent/task_test.go
@@ -1169,21 +1169,6 @@ func TestSubSinkForwardsUsageToParent(t *testing.T) {
 	}
 }
 
-func TestSubSinkForwardsWorkspaceMutationToParent(t *testing.T) {
-	parent := newWorkspaceSignalSink()
-	event.RecordWorkspaceMutation(subSinkFor("task_1", parent), event.WorkspaceMutation{
-		ToolID: "write", ToolName: "write_file", Paths: []string{"child.go"}, Content: true,
-	})
-	select {
-	case mutation := <-parent.mutations:
-		if mutation.ToolName != "write_file" || len(mutation.Paths) != 1 || mutation.Paths[0] != "child.go" {
-			t.Fatalf("forwarded workspace mutation = %+v", mutation)
-		}
-	default:
-		t.Fatal("sub-agent workspace mutation was not forwarded")
-	}
-}
-
 func TestTaskToolCarriesRecentKeepIntoSubsessions(t *testing.T) {
 	task := NewTaskTool(&mockProvider{name: "sub"}, nil, tool.NewRegistry(), 20, 0, 7, 0, 0, 0, 0, 0.0, "", "sys", nil, 0, "", "", nil)
 	if task.recentKeep != 7 {
